Move capture device interaction into a subscription
The main (UI) thread no longer handles the device connection. Instead, we
communicate by passing messages to the thread managing the device. The UI
thread gets the Send half of a channel as the first messages and can
subsequently pass Request instances. Responses are sent by the subscription
thread as events which are asynchronously received by the UI thread in the
subscription() method.

This rework is part of an effort to bring up live image streaming while being
able to change controls and formats concurrently.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>